optimisation: using Float instead of Double

+++ examples/GridworldStep.hs
@@ -206,10 +206,10 @@ modelBuilder colOut =
   where inpLen = genericLength (netInp initState)
 
 
-netInp :: St -> [Double]
+netInp :: St -> [Float]
 netInp st = [scaleNegPosOne (0, fromIntegral maxX) $ fromIntegral $ fst (getCurrentIdx st), scaleNegPosOne (0, fromIntegral maxY) $ fromIntegral $ snd (getCurrentIdx st)]
 
-tblInp :: St -> [Double]
+tblInp :: St -> [Float]
 tblInp st = [fromIntegral $ fst (getCurrentIdx st), fromIntegral $ snd (getCurrentIdx st)]
 
 names = ["random", "up   ", "down ", "left ", "right"]
@@ -259,7 +259,7 @@ goalState :: (AgentType -> St -> IO (Reward St, St, EpisodeEnd)) -> AgentType ->
 goalState f tp st = do
   x <- randomRIO (0, maxX :: Int)
   y <- randomRIO (0, maxY :: Int)
-  r <- randomRIO (0, 8 :: Double)
+  r <- randomRIO (0, 8 :: Float)
   let stepRew (Reward re, s, e) = (Reward $ re + r, s, e)
   case getCurrentIdx st of
     (x', y')
@@ -298,7 +298,7 @@ fromIdx (m,n) = St $ zipWith (\nr xs -> zipWith (\nr' ys -> if m == nr && n == n
   where base = replicate 5 [0,0,0,0,0]
 
 
-allStateInputs :: M.Map [Double] St
+allStateInputs :: M.Map [Float] St
 allStateInputs = M.fromList $ zip (map netInp [minBound..maxBound]) [minBound..maxBound]
 
 mInverseSt :: Maybe (NetInputWoAction -> Maybe (Either String St))
